auth added to user controller

+++ task-manager/src/user/user.controller.ts
@@ -9,24 +9,40 @@ import {
   ConflictException,
   NotFoundException,
   HttpCode,
+  UseGuards,
+  Request,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { AuthGuard } from 'src/auth/auth.guard';
+import { IsNull } from 'typeorm';
 
 @Controller('user')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
+  @UseGuards(AuthGuard)
   @Get(':id')
-  async findOne(@Param('id', ParseIntPipe) id: number) {
-    return await this.userService.findById(+id);
+  async findOne(@Param('id', ParseIntPipe) id: number, @Request() req) {
+    const userId = req.userPayload.id;
+    if (userId !== id) throw new ForbiddenException();
+    const user = await this.userService.findById(+id);
+    if (!user) throw new NotFoundException('user not found.');
+    return user;
   }
 
+  @UseGuards(AuthGuard)
   @Patch(':id')
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateUserDto: UpdateUserDto,
+    @Request() req,
   ) {
+    const user = await this.userService.findById(+id);
+    if (!user) throw new NotFoundException('user not found.');
+    const userId = req.userPayload.id;
+    if (userId !== id) throw new ForbiddenException();
     if (!!updateUserDto.email) {
       const existingUser = await this.userService.findByEmail(
         updateUserDto.email,
@@ -42,13 +58,15 @@ export class UserController {
     return await this.userService.update(id, updateUserDto);
   }
 
+  @UseGuards(AuthGuard)
   @Delete(':id')
   @HttpCode(204)
-  async remove(@Param('id', ParseIntPipe) id: number) {
-    const result = await this.userService.remove(id);
-    if (result.affected == 0) {
-      throw new NotFoundException('User with given id does not exist.');
-    }
-    return null;
+  async remove(@Param('id', ParseIntPipe) id: number, @Request() req) {
+    const user = await this.userService.findById(+id);
+    if (!user) throw new NotFoundException('user not found.');
+    const userId = req.userPayload.id;
+    if (userId !== id) throw new ForbiddenException();
+    await this.userService.remove(id);
+    return IsNull;
   }
 }
